feat(infinite-canvas): make buffer ref consistant
Update dependency resize-canvas which now adjusts the buffer canvas in
place, so reference will stay the same once you get it. Also, add
getOrigin which was missing per documentation. In addition, the instance
is frozen. Of note: the whole package was refactored to use the revealing
module pattern.

BREAKING CHANGE: The instance object now is returned frozen, so it can't
be modified. This is different then previous behavior.